+++ test/unit/exports.test.js
@@ -74,15 +74,6 @@ describe("Exports", function () {
     });
 
     describe("Passwordless", function () {
-        it("Utils", function () {
-            let _default = require("../../recipe/passwordless/utils");
-            let { consumeCode, createCode, resendCode } = require("../../recipe/passwordless/utils");
-
-            assert(createCode !== undefined && _default.createCode !== undefined);
-            assert(consumeCode !== undefined && _default.consumeCode !== undefined);
-            assert(resendCode !== undefined && _default.resendCode !== undefined);
-        });
-
         it("Index", function () {
             let _default = require("../../recipe/passwordless");
             let {